move user_can_change_project to conftest.py
